fix!: drop Python 3.8 support
Closes #21.

I am not testing pypara on Python 3.8 as Nix fails to build the
development and testing environment for Python 3.8. I can bypass Nix
for tests or try to make it work, but I do not think that it is worth
the effort.

Also, I anticipate issues with zoneinfo that is backported to Python
<3.9 via backports.zoneinfo library.

Therefore, I am dropping support for Python 3.8.